feat: expose proving service as lib
