+++ bin/proving-service/src/commands/mod.rs
@@ -1,10 +1,9 @@
 use clap::Parser;
+use miden_proving_service::api::{MIDEN_PROVING_SERVICE, ProverType};
 use proxy::StartProxy;
 use tracing::instrument;
 use update_workers::{AddWorkers, RemoveWorkers, UpdateWorkers};
-use worker::{ProverType, StartWorker};
-
-use crate::utils::MIDEN_PROVING_SERVICE;
+use worker::StartWorker;
 
 pub mod proxy;
 pub mod update_workers;
