+++ bin/proving-service/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod api;
+pub mod error;
+pub mod generated;
